refactoring : Gathering Refacotring
